MERGE : 배포

See merge request s10-webmobile2-sub2/S10P12A609!13

+++ seas-be/src/main/java/com/ssafy/seas/common/entity/BaseEntity.java
@@ -3,13 +3,13 @@ package com.ssafy.seas.common.entity;
 import java.time.LocalDateTime;
 
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
 
@@ -19,7 +19,7 @@ import lombok.Getter;
 public class BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 
 	@CreatedDate
 	private LocalDateTime createdAt;
